// sdk/src/gateway/generated-client/models/GatewayCreateOrderOneOf.ts
/* tslint:disable */
/* eslint-disable */
/**
 * gateway-api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { GatewayCreateOrderOneOfOnramp } from './GatewayCreateOrderOneOfOnramp';
import {
    GatewayCreateOrderOneOfOnrampFromJSON,
    GatewayCreateOrderOneOfOnrampFromJSONTyped,
    GatewayCreateOrderOneOfOnrampToJSON,
    GatewayCreateOrderOneOfOnrampToJSONTyped,
} from './GatewayCreateOrderOneOfOnramp';

/**
 * 
 * @export
 * @interface GatewayCreateOrderOneOf
 */
export interface GatewayCreateOrderOneOf {
    /**
     * 
     * @type {GatewayCreateOrderOneOfOnramp}
     * @memberof GatewayCreateOrderOneOf
     */
    onramp: GatewayCreateOrderOneOfOnramp;
}

/**
 * Check if a given object implements the GatewayCreateOrderOneOf interface.
 */
export function instanceOfGatewayCreateOrderOneOf(value: object): value is GatewayCreateOrderOneOf {
    if (!('onramp' in value) || value['onramp'] === undefined) return false;
    return true;
}

export function GatewayCreateOrderOneOfFromJSON(json: any): GatewayCreateOrderOneOf {
    return GatewayCreateOrderOneOfFromJSONTyped(json, false);
}

export function GatewayCreateOrderOneOfFromJSONTyped(json: any, ignoreDiscriminator: boolean): GatewayCreateOrderOneOf {
    if (json == null) {
        return json;
    }
    return {
        
        'onramp': GatewayCreateOrderOneOfOnrampFromJSON(json['onramp']),
    };
}

export function GatewayCreateOrderOneOfToJSON(json: any): GatewayCreateOrderOneOf {
    return GatewayCreateOrderOneOfToJSONTyped(json, false);
}

export function GatewayCreateOrderOneOfToJSONTyped(value?: GatewayCreateOrderOneOf | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'onramp': GatewayCreateOrderOneOfOnrampToJSON(value['onramp']),
    };
}