// sdk/src/gateway/generated-client/models/GatewayMaxSpendable.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BOB Gateway API
 * BOB Gateway API for Bitcoin swaps
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { GatewayTokenAmount } from './GatewayTokenAmount';
import {
    GatewayTokenAmountFromJSON,
    GatewayTokenAmountFromJSONTyped,
    GatewayTokenAmountToJSON,
    GatewayTokenAmountToJSONTyped,
} from './GatewayTokenAmount';

/**
 * 
 * @export
 * @interface GatewayMaxSpendable
 */
export interface GatewayMaxSpendable {
    /**
     * 
     * @type {GatewayTokenAmount}
     * @memberof GatewayMaxSpendable
     */
    amount: GatewayTokenAmount;
    /**
     * 
     * @type {string}
     * @memberof GatewayMaxSpendable
     */
    userAddress: string;
}

/**
 * Check if a given object implements the GatewayMaxSpendable interface.
 */
export function instanceOfGatewayMaxSpendable(value: object): value is GatewayMaxSpendable {
    if (!('amount' in value) || value['amount'] === undefined) return false;
    if (!('userAddress' in value) || value['userAddress'] === undefined) return false;
    return true;
}

export function GatewayMaxSpendableFromJSON(json: any): GatewayMaxSpendable {
    return GatewayMaxSpendableFromJSONTyped(json, false);
}

export function GatewayMaxSpendableFromJSONTyped(json: any, ignoreDiscriminator: boolean): GatewayMaxSpendable {
    if (json == null) {
        return json;
    }
    return {
        
        'amount': GatewayTokenAmountFromJSON(json['amount']),
        'userAddress': json['userAddress'],
    };
}

export function GatewayMaxSpendableToJSON(json: any): GatewayMaxSpendable {
    return GatewayMaxSpendableToJSONTyped(json, false);
}

export function GatewayMaxSpendableToJSONTyped(value?: GatewayMaxSpendable | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'amount': GatewayTokenAmountToJSON(value['amount']),
        'userAddress': value['userAddress'],
    };
}