// sdk/src/gateway/generated-client/models/GatewayOrderInfoV2.ts
export interface GatewayOrderInfoV2 {
    /**
     * 
     * @type {string}
     * @memberof GatewayOrderInfoV2
     */
    depositAddress?: string | null;
    /**
     * 
     * @type {OrderDestinationInfo}
     * @memberof GatewayOrderInfoV2
     */
    dstInfo: OrderDestinationInfo;
    /**
     * 
     * @type {number}
     * @memberof GatewayOrderInfoV2
     */
    estimatedTimeInSecs?: number | null;
    /**
     * 
     * @type {string}
     * @memberof GatewayOrderInfoV2
     */
    id: string;
    /**
     * 
     * @type {ChainTxInfo}
     * @memberof GatewayOrderInfoV2
     */
    srcInfo: ChainTxInfo;
    /**
     * 
     * @type {GatewayOrderStatusV2}
     * @memberof GatewayOrderInfoV2
     */
    status: GatewayOrderStatusV2;
    /**
     * 
     * @type {number}
     * @memberof GatewayOrderInfoV2
     */
    timestamp: number;
}